mt5_gateway/requirements.txt
fastapi
uvicorn[standard]
MetaTrader5
pydantic
python-dotenv